+++ examples/composable-example/src/main/java/com/accenture/demo/tasks/EncryptFields.java
@@ -16,7 +16,7 @@
 
  */
 
- package com.accenture.demo.tasks;
+package com.accenture.demo.tasks;
 
 import org.platformlambda.core.annotations.PreLoad;
 import org.platformlambda.core.models.TypedLambdaFunction;