+++ system/platform-core/src/main/java/org/platformlambda/core/logging/JsonAppender.java
@@ -19,8 +19,6 @@
 package org.platformlambda.core.logging;
 
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.ToNumberPolicy;
 import org.apache.logging.log4j.core.*;
 import org.apache.logging.log4j.core.appender.AbstractAppender;
 import org.apache.logging.log4j.core.config.Property;