+++ system/platform-core/src/main/java/org/platformlambda/core/services/DistributedTrace.java
@@ -19,12 +19,11 @@
 package org.platformlambda.core.services;
 
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.ToNumberPolicy;
 import org.platformlambda.core.annotations.EventInterceptor;
 import org.platformlambda.core.annotations.ZeroTracing;
 import org.platformlambda.core.models.EventEnvelope;
 import org.platformlambda.core.models.TypedLambdaFunction;
+import org.platformlambda.core.serializers.SimpleMapper;
 import org.platformlambda.core.system.Platform;
 import org.platformlambda.core.system.EventEmitter;
 import org.slf4j.Logger;
@@ -39,8 +38,7 @@ import java.util.Map;
 @ZeroTracing
 public class DistributedTrace implements TypedLambdaFunction<EventEnvelope, Void> {
     private static final Logger log = LoggerFactory.getLogger(DistributedTrace.class);
-    private static final Gson serializer = new GsonBuilder().disableHtmlEscaping()
-                                                .setObjectToNumberStrategy(ToNumberPolicy.LONG_OR_DOUBLE).create();
+    private static final Gson serializer = SimpleMapper.getInstance().getPrettyGson();
     private static final String DISTRIBUTED_TRACE_FORWARDER = "distributed.trace.forwarder";
     private static final String TRANSACTION_JOURNAL_RECORDER = "transaction.journal.recorder";
     private static final String TRACE = "trace";